refactor(code-block): extract code block logic into hooks
